Refactor `effects.rs` into effects module.
